fix: itoa bad declaration fixed

+++ kfs_3/srcs/libs/stdlib/itoa.c
@@ -2,7 +2,7 @@
 
 static const char *base = "0123456789abcdefghijklmnopqrstuvwxyz";
 
-char* itoa (int val, char * buffer, int radix) {
+char *itoa(unsigned long val, char *buffer, unsigned long radix) {
     int next = val / radix;
     if (next) buffer = itoa(next, buffer, radix) + 1;
     buffer[0] = base[val % radix];
